+++ .gitignore
@@ -11,4 +11,7 @@ ollama
 test_data
 *.crt
 llm/build
+build/*/*/*
+!build/**/placeholder
+llama/build
 __debug_bin*
